+++ featuretools/primitives/standard/transform/exponential/exponential_weighted_variance.py
@@ -44,6 +44,7 @@ class ExponentialWeightedVariance(TransformPrimitive):
     name = "exponential_weighted_variance"
     input_types = [ColumnSchema(semantic_tags={"numeric"})]
     return_type = ColumnSchema(logical_type=Double, semantic_tags={"numeric"})
+    uses_full_dataframe = True
 
     def __init__(self, com=None, span=None, halflife=None, alpha=None, ignore_na=False):
         if all(x is None for x in [com, span, halflife, alpha]):