+++ featuretools/primitives/standard/transform/time_series/rolling_outlier_count.py
@@ -90,6 +90,7 @@ class RollingOutlierCount(TransformPrimitive):
         ColumnSchema(semantic_tags={"numeric"}),
     ]
     return_type = ColumnSchema(logical_type=Double, semantic_tags={"numeric"})
+    uses_full_dataframe = True
 
     def __init__(self, window_length=3, gap=1, min_periods=0):
         self.window_length = window_length